shared classes moved to types
